requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
backoff==1.11.1
Brotli==1.1.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
distlib==0.3.8
et-xmlfile==1.1.0
ffmpy==0.3.2
filelock==3.13.1
frozenlist==1.4.1
identify==2.5.35
idna==3.6
multidict==6.0.5
mutagen==1.47.0
nodeenv==1.8.0
openpyxl==3.1.2
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.2
pycryptodomex==3.20.0
pydub==0.25.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-file==2.0.0
six==1.16.0
spotify==0.10.2
spotipy==2.23.0
tldextract==5.1.1
urllib3==2.2.1
validators==0.22.0
virtualenv==20.25.1
websockets==12.0
yarl==1.9.4
yt-dlp==2024.3.10
